requirements.txt
flask-script
flask-bootstrap
flask-wtf